refactor(cli): split cmd/cli/show.go by feature family (#4660)
cmd/cli/show.go had grown to ~2100 LOC as the remote-CLI `show`
umbrella: the top-level dispatcher plus the security, flow, NAT,
interfaces, protocols, system, services, and DHCP renderers all in one
file. That made local-vs-remote parity review — the whole point of a
remote CLI that must stay bit-identical to the daemon-local surface —
hard to reason about per feature.

Split the per-feature handlers into cohesive sibling files in the same
package (unexported symbols stay reachable):

  show_security.go   zones/policies/screen/match-policies/statistics/
                     log/vrrp/ike/ipsec renderers (654 LOC)
  show_flow.go       flow session parse + render + summary + flow
                     statistics (395 LOC)
  show_nat.go        source/destination NAT renderers (298 LOC)
  show_interfaces.go show interfaces (47 LOC)
  show_protocols.go  ospf/bgp/bfd/rip/isis (85 LOC)
  show_system.go     commit/rollback/uptime/... (137 LOC)
  show_services.go   rpm/ip-monitoring/app-id/ddns (51 LOC)
  show_dhcp.go       leases / client-identifier (62 LOC)

show.go (now 449 LOC) keeps the handleShow dispatcher, the config-mode
handleConfigShow, showRoutes, and the shared text-proxy helpers
(showText / showTextFiltered / showSystemInfo). The gRPC call sequences
and the text-proxy fallthrough are preserved verbatim, so the remote
CLI output is unchanged.

This is pure code motion — no behavior change. All 44 top-level
declarations were extracted verbatim (go/ast, doc-comment-inclusive
spans) and verified byte-for-byte against origin/master: 44 decls on
both sides, 0 missing, 0 extra, 0 body diffs. go build ./... clean,
go vet ./cmd/cli/... unchanged (the pre-existing monitor.go lock-copy
warning is on origin/master too), go test ./cmd/cli/... green. No new
test, since nothing changed at runtime.
