+++ cmd/cli/show_services.go
@@ -1,0 +1,51 @@
+package main
+
+import (
+	"fmt"
+)
+
+func (c *ctl) handleShowServices(args []string) error {
+	if len(args) == 0 {
+		printRemoteTreeHelp("show services:", "show", "services")
+		return nil
+	}
+	switch args[0] {
+	case "rpm":
+		return c.showText("rpm")
+	case "ip-monitoring":
+		// #1827: only `ip-monitoring status` is valid per cmdtree.
+		rest := args[1:]
+		if len(rest) == 0 {
+			printRemoteTreeHelp("show services ip-monitoring:", "show", "services", "ip-monitoring")
+			return nil
+		}
+		if rest[0] != "status" {
+			return fmt.Errorf("unknown ip-monitoring target: %s (expected `status`)", rest[0])
+		}
+		return c.showText("services-ip-monitoring-status")
+	case "application-identification":
+		// #653: surface what xpf AppID actually does today. Per
+		// cmdtree the only valid leaf is `application-identification
+		// status`; reject anything else so typos surface as usage
+		// errors instead of being silently swallowed.
+		rest := args[1:]
+		if len(rest) == 0 {
+			printRemoteTreeHelp("show services application-identification:",
+				"show", "services", "application-identification")
+			return nil
+		}
+		if rest[0] != "status" {
+			return fmt.Errorf("unknown application-identification target: %s "+
+				"(expected `status`)", rest[0])
+		}
+		return c.showText("application-identification-status")
+	case "dynamic-dns":
+		// #2691 P2: Surface A (router/interface-address) DDNS status.
+		if len(args) >= 2 && args[1] == "detail" {
+			return c.showText("services-dynamic-dns-detail")
+		}
+		return c.showText("services-dynamic-dns")
+	default:
+		return fmt.Errorf("unknown services target: %s", args[0])
+	}
+}
