+++ cmd/cli/show_system.go
@@ -1,0 +1,137 @@
+package main
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	pb "github.com/psaab/xpf/pkg/grpcapi/xpfv1"
+)
+
+func (c *ctl) handleShowSystem(args []string) error {
+	if len(args) == 0 {
+		printRemoteTreeHelp("show system:", "show", "system")
+		return nil
+	}
+
+	switch args[0] {
+	case "commit":
+		if len(args) >= 2 && args[1] == "history" {
+			return c.showText("commit-history")
+		}
+		printRemoteTreeHelp("show system commit:", "show", "system", "commit")
+		return nil
+
+	case "rollback":
+		if len(args) >= 2 {
+			if args[1] == "compare" && len(args) >= 3 {
+				n, err := strconv.Atoi(args[2])
+				if err != nil || n < 1 {
+					return fmt.Errorf("usage: show system rollback compare <N>")
+				}
+				resp, err := c.client.ShowCompare(c.ctx(), &pb.ShowCompareRequest{
+					RollbackN: int32(n),
+				})
+				if err != nil {
+					return fmt.Errorf("%v", err)
+				}
+				if resp.Output == "" {
+					fmt.Println("No differences found")
+				} else {
+					fmt.Print(resp.Output)
+				}
+				return nil
+			}
+
+			n, err := strconv.Atoi(args[1])
+			if err != nil || n < 1 {
+				return fmt.Errorf("usage: show system rollback <N>")
+			}
+			format := pb.ConfigFormat_HIERARCHICAL
+			rest := strings.Join(args[2:], " ")
+			if strings.Contains(rest, "| display set") {
+				format = pb.ConfigFormat_SET
+			} else if strings.Contains(rest, "| display xml") {
+				format = pb.ConfigFormat_XML
+			} else if strings.Contains(rest, "compare") {
+				resp, err := c.client.ShowCompare(c.ctx(), &pb.ShowCompareRequest{
+					RollbackN: int32(n),
+				})
+				if err != nil {
+					return fmt.Errorf("%v", err)
+				}
+				if resp.Output == "" {
+					fmt.Println("No differences found")
+				} else {
+					fmt.Print(resp.Output)
+				}
+				return nil
+			}
+			resp, err := c.client.ShowRollback(c.ctx(), &pb.ShowRollbackRequest{
+				N:      int32(n),
+				Format: format,
+			})
+			if err != nil {
+				return fmt.Errorf("%v", err)
+			}
+			fmt.Print(resp.Output)
+			return nil
+		}
+
+		resp, err := c.client.ListHistory(c.ctx(), &pb.ListHistoryRequest{})
+		if err != nil {
+			return fmt.Errorf("%v", err)
+		}
+		if len(resp.Entries) == 0 {
+			fmt.Println("No rollback history available")
+			return nil
+		}
+		for _, e := range resp.Entries {
+			fmt.Printf("  rollback %d: %s\n", e.Index, e.Timestamp)
+		}
+		return nil
+
+	case "uptime":
+		return c.showSystemInfo("uptime")
+	case "memory":
+		return c.showSystemInfo("memory")
+	case "storage":
+		return c.showText("storage")
+	case "processes":
+		return c.showSystemInfo("processes")
+	case "alarms":
+		return c.showText("alarms")
+	case "users":
+		return c.showSystemInfo("users")
+	case "connections":
+		return c.showSystemInfo("connections")
+	case "license":
+		fmt.Println("License: open-source (no license required)")
+		return nil
+	case "services":
+		return c.showText("system-services")
+	case "ntp":
+		return c.showText("ntp")
+	case "login":
+		return c.showText("login")
+	case "syslog":
+		return c.showText("system-syslog")
+	case "internet-options":
+		return c.showText("internet-options")
+	case "root-authentication":
+		return c.showText("root-authentication")
+	case "backup-router":
+		return c.showText("backup-router")
+	case "buffers":
+		if len(args) >= 2 && args[1] == "detail" {
+			return c.showText("buffers-detail")
+		}
+		return c.showText("buffers")
+	case "boot-messages":
+		return c.showSystemInfo("boot-messages")
+	case "core-dumps":
+		return c.showText("core-dumps")
+	default:
+		return fmt.Errorf("unknown show system target: %s", args[0])
+	}
+}
